requirements.txt
pyyaml
requests
influxdb-client[ciso]
yamllint
